blockchain/requirements.txt
pyteal>=0.20.1
py-algorand-sdk>=2.4.0
python-dotenv>=1.0.0
pytest>=7.4.0
